+++ CHANGELOG.md
@@ -78,6 +78,7 @@ Currently the versioning policy of this project follows [Semantic Versioning v2.
   - `toString(Object[])` method from `edu.umd.cs.findbugs.util.Strings` deprecated since 2010,
   - `closeSilently(OutputStream)` method from `edu.umd.cs.findbugs.util.Util` deprecated since 2018,
   - `closeSilently(Closeable)` method from `edu.umd.cs.findbugs.util.Util` deprecated since 2018,
+  - `closeSilently(ZipFile)` method from `edu.umd.cs.findbugs.util.Util` deprecated since 2018,
 - Removed old deprecated fields:
   - String `RELEASE` from `edu.umd.cs.findbugs.Version` deprecated since 2018,